5033 Move SystemTestDataGenerationJobStore to config module
